% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tdi.R
\name{plot.tdi}
\alias{plot.tdi}
\title{Bland-Altman plot}
\usage{
\method{plot}{tdi}(x, main = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{input object of class \code{tdi} resulting from a call of the function \code{\link[TDIagree]{TDI}}.}

\item{main}{overall title for the plot (to be passed to \code{main} argument in \code{\link[base]{plot}}).}

\item{xlim}{the \eqn{x} limits of the plot (to be passed to \code{xlim} argument in \code{\link[base]{plot}}). \cr
The default value, NULL, indicates that the range of the mean values should be used.}

\item{ylim}{the \eqn{y} limits of the plot (to be passed to \code{ylim} argument in \code{\link[base]{plot}}). \cr
The default value, NULL, indicates that the range of the differences values should be used.}

\item{...}{other graphical parameters (to be passed to \code{\link[base]{plot}}).}
}
\value{
A Bland-Altman plot of the data in \code{x}.
}
\description{
This function creates a Bland-Altman plot, which is used to evaluate the agreement among the quantitative measures taken by two raters.
The plot displays the mean of the measurements from both raters in the x-axis and the differences between the measures taken by the two raters in the y-axis.
}
\section{Note}{

Future versions of the package will include the TDI and UB estimates in the plot.
}

\examples{
set.seed(2025)

n <- 100
y_A1 <- rnorm(n, 50, 10) # rater A, replicate 1
y_A2 <- rnorm(n, 50, 10) # rater A, replicate 2
y_B1 <- rnorm(n, 30, 15) # rater B, replicate 1
y_B2 <- rnorm(n, 30, 15) # rater B, replicate 2

ex_data <- data.frame(y = c(y_A1, y_A2, y_B1, y_B2), rater = factor(rep(c("A", "B"), each = 2*n)),
                      replicate = factor(rep(rep(1:2, each = n), 2)), subj = factor(rep(1:n, 4)))

tdi <- TDI(ex_data, y, subj, rater, replicate, p = c(0.8, 0.9),
           method = c("Choudhary P", "Escaramis", "Choudhary NP", "Perez-Jaume and Carrasco"))
plot(tdi)

# enhance plot
par(las = 1, font.lab = 2, font.axis = 2)
plot(tdi, xlim = c(10, 70), ylim = c(-60, 80), pch = 16, col = "red")

}
\references{
Altman, D. G., & Bland, J. M. (1983). Measurement in medicine: the analysis of method comparison studies. Journal of the Royal Statistical Society Series D: The Statistician, 32(3), 307-317.
}
