% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIagree.R
\docType{package}
\name{TDIagree-package}
\alias{TDIagree}
\alias{TDIagree-package}
\title{Assessment of agreement for quantitative normal and non-normal data using the total deviation index}
\description{
The total deviation index (TDI) is an unscaled statistical measure used to evaluate the deviation between paired quantitative measurements when assessing
the extent of agreement between different raters. It describes a boundary such that a large specified proportion of the differences in paired measurements
are within the boundary (Lin, 2000) \url{https://doi.org/10.1002/(SICI)1097-0258(20000130)19:2<255::AID-SIM293>3.0.CO;2-8}. This R package implements some methodologies
existing in the literature for TDI estimation and inference in the case of two raters.
}
\section{Functions}{

\code{\link[TDIagree]{TDI}}
}

\section{Methods}{

\code{\link[TDIagree]{print.tdi}}, \code{\link[TDIagree]{plot.tdi}}
}

\references{
Lin, L. I. K. (2000). Total deviation index for measuring individual agreement with applications in laboratory performance and bioequivalence. Statistics in medicine, 19.2:255-270.
}
\author{
\strong{Maintainer}: Anna Felip-Badia \email{annafelipibadia@gmail.com}

Authors:
\itemize{
  \item Sara Perez-Jaume
}

}
